fix(Tests): move stub classes into their own files
