+++ tests/Stubs/AuditableModel.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace OwenIt\Auditing\Tests\Stubs;
+
+use Illuminate\Database\Eloquent\Model;
+use OwenIt\Auditing\Auditable;
+use OwenIt\Auditing\Contracts\Auditable as AuditableContract;
+
+class AuditableModel extends Model implements AuditableContract
+{
+    use Auditable;
+}
